==> 02 Scripts/1.0 Importing libraries and Python data types.ipynb <==
# 01. Importing libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# 02. Working with pandas data types

x = 40
y = 60

z = x + y
print(z)

a = x + x + 20
print(z)

b = 140 - x
print(b)

# float operations

# float operations
c = 5.25
d = 10.50

print(c/d)

print(d/c)

# Concatenation

a = 'wor'
b = 'ld'
print(a+b)

# concatinating two word strings
a = 'Hello, '
b = 'World!'

print(a+b)

a = 'importing libraries '
b = 'and Python data types'
print(a+b)
