==> 02 Scripts/2.0 loading data set into jupyter notebook.ipynb <==
# importing libraries

import pandas as pd
import numpy as np
import os

# set the folder path to variable
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# creating list of columns excluding "eval_set"
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# reading data

df = pd.read_csv(os.path.join(path, "02 Data", "Original Data", "orders.csv"), usecols = vars_list)

# printing first 5 rows of data
df.head()

# products data

df_prods = pd.read_csv(os.path.join(path, "02 Data", "Original Data", "products.csv"))

# printing first 20 rows
df_prods.head(20)

#printing last 35 rows
df_prods.tail(35)

# printing columns names
df_prods.columns

# printing number of rows and columns
df_prods.shape

# products dataframe has 49693 rows and 5 columns

df_prods.describe()

# the maximum value of "aisle_id " is 134

df_prods.dtypes

# the data type of "department_id " is int64
