==> 02 Scripts/3.0 data wrangling and subsetting.ipynb <==
# importing libraries

import pandas as pd
import numpy as np
import os

# set the folder path to variable
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

#orders data
df_ords = pd.read_csv(os.path.join(path, "02 Data", "Original Data", "orders.csv"), index_col = False)
#products data
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'products.csv'), index_col = False)
#departments data
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original data', 'departments.csv'), index_col = False)

# data wrangling

# dropping eval_set from data
df_ords = df_ords.drop(columns = ['eval_set'])

# changing column name of order_dow
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# changing data types of user id and order id 
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

# department data set

df_dep_t = df_dep.T

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

# removing the very first row, in other words using data after the 1st row as the 1st will be taken as headers of the data set
df_dep_t_new = df_dep_t[1:]

# settig 'new_header' as the column names
df_dep_t_new.columns = new_header

df_dep_t_new

data_dict = df_dep_t_new.to_dict('index')

# step 2: it doesn't make sense to make stats on the order number and user id so i have changed the data types to string as well
df_ords['order_number'] = df_ords['order_number'].astype('str')

df_ords.dtypes

#step 3: i only found order_dow which i have already changed to a suitable format

df_ords.head()

# step 4:  finding the busiest hour
df_ords['order_hour_of_day'].value_counts()

# the busiest hour of the day is 10 am followed by 11 am

# step 5:
data_dict.get('4')

The department_id "4" corresponds to the category "produce"

# creating subset for breakfast items
data_dict

# step 6: creating the subset for breakfast
df_prods_breakfast = df_prods.loc[df_prods['department_id'] == 14]

df_prods_breakfast

df_prods.head()

#  items from the following departments: alcohol, deli, beverages, and meat/seafood
data_dict

# step. 7: creating subset for dinner party.. sales teams can use this data frame for further analysis
df_prods_dinner_party = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]

# step 8: total rows of last data frame:
df_prods_dinner_party.shape

it has 7650 rows

#step 9: i have created a df to know about the stats of this user
df_user1 = df_ords[df_ords['user_id']=='1']

df_user1

# to get the stats for df_user1
df_user1.describe().round(2)

# observation:
This user places orders exclusively from Sunday to Wednesday, avoiding any orders from Thursday to Saturday. 
Over half the time, their orders are made at or before 8 a.m. 
On average, they place an order every 19 days. They have never gone more than 30 days without placing an order and have placed two orders on the same day on at least one occasion.

# exporting data frames

# orders data frame
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'), index = False)

# department data frame
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'), index = False)
