==> 02 Scripts/4.0 Data Consistency Checks.ipynb <==
# importing libraries
import pandas as pd
import numpy as np
import os

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_ords.head()

df_test = pd.DataFrame()  #creates a new dataframe called df_test. and then putting mixed type values to the column mix
df_test['mix'] = ['a', 'b', 1, True]

df_test.dtypes

# check for mix data types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# updating datatype to str
df_test['mix'] = df_test['mix'].astype('str')

# running stats on df_prods
df_prods.describe()

# finding median of prices column
df_prods['prices'].median()

# finding mean of prices column
df_prods['prices'].mean()

# finding outliers in prices column
outliers = df_prods[df_prods['prices'] > df_prods['prices'].mean() + (2 * df_prods['prices'].std())]

outliers

## handling outliers
#### because of the 2 extreme outliers the mean value is affected, Replacing the outliers values in the prices column with the meadian value of 7.1. makes more sense

 # replacing the outliers with median value of 7.1
df_prods = df_prods.replace({"prices":{14900.0:7.1, 99999.0:7.1}})

df_prods.describe()

### Observation 
#### After conducting exploratory analysis on the df_products DataFrame, outliers were identified in the prices column. The column had a mean of 9.99 and a standard deviation of 453.6. This unusually high standard deviation, more than 2 standard deviations away from the mean, highlighted the presence of outlier values.
#### Specifically, two entries were identified as outliers: prices of 14,900 and 99,999. These outliers were replaced with the median value of 7.1. Replacing these extreme values with the median is recommended because the minimum value of 1 and the maximum value of 99,999 create a wide range, making the median a more robust measure compared to the mean, which is sensitive to outliers.
#### After replacing the outliers, the mean of the prices column decreased to 7.6, reflecting a more accurate representation of the data.

# finding missing values in products df
df_prods.isnull().sum()

# creating new data frame containing only null values in product name
df_nan = df_prods[df_prods['product_name'].isnull()== True]

df_nan

# creating a new clean data fram that conatins no null values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

df_prods.shape

# finding duplicates in the data frame
df_dup = df_prods_clean[df_prods_clean.duplicated()]

df_dup

# creating a new data frame without duplicates
df_prods_clean_no_dup =  df_prods_clean.drop_duplicates()

df_prods_clean_no_dup.shape

# df_ords

# step 2:
df_ords.describe()

#### After examining the values in each column, no unreasonable values were found

# step 3: 
# checking for mixed type in orders data set
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)


####  step 4:
#### The `df_ords` DataFrame contains no mixed-type data in any of its columns or variables. This was verified by using a function to search for mixed-type data within the columns, which returned a result of NULL.

# step 5: checking for missing values in df_ords
df_ords.isnull().sum()

df_ords[df_ords['days_since_prior_order'].isnull() == True]

#### step 6: 
#### after examinig the data it was found that, The `days_since_prior_order` column in the original `df_ords` DataFrame contains a total of 206,209 missing values. Each unique `user_id` has exactly one missing value in this column, which corresponds to their very first order. This is logical, as there is no prior order for the first purchase of a user. Therefore, it is unnecessary to remove these missing values.

# step 7:
# checking for duplicates in the df_ords
df_ords[df_ords.duplicated()]

 #### step 8: 
 #### there was found to be no duplicates in the entire df_ords data frame

# step 9:
# exporting the data sets
df_prods_clean_no_dup.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index = False)

# exporting the ords data sets
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index = False)
