==> 02 Scripts/4.6_combining_data_demo.ipynb <==
## 4.6 Merging and exporting data

### This script contains the following points:

#### 1. Create data to experiment on
#### 2. Concatenate dataframes
#### 3. Append data
#### 4. Merge data

# Import libraries

import pandas as pd
import numpy as np
import os

#### 1. Create data to experiment on

# Define a dictionary containing January 2020 data 

data1 = {'customer_id':['6732', '767', '890', '635'], 
        'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'], 
        'purchased_meat':[0, 13, 3, 4], 
        'purchased_alcohol':[1, 2, 10, 0],
        'purchased_snacks': [10, 5, 1, 7]} 

# Define a dictionary containing February 2020 data 

data2 = {'customer_id':['6732', '767', '890', '635'], 
        'month':['Feb-20', 'Feb-20', 'Feb-20', 'Feb-20'], 
        'purchased_meat':[0, 10, 5, 3], 
        'purchased_alcohol':[2, 4, 14, 0],
        'purchased_snacks': [15, 3, 2, 6]} 

# Convert the dictionary into DataFrame  
df = pd.DataFrame(data1,index=[0, 1, 2, 3])
df_1 = pd.DataFrame(data2,index=[0, 1, 2, 3])

df

df_1

#### 2. Concatenate dataframes

# Create a list that contains our dataframes

frames = [df, df_1]

# Check the output

frames

# Check the data types to be sure it is a list

type(frames)

# Concatenate the dataframes using default options

df_concat = pd.concat(frames)

# Check the output

df_concat

# Concatenate the dataframes using the axis = 1 --> create a wide format as an output

df_concat = pd.concat(frames, axis = 1)

# Check the output

df_concat

#### 3. Append data

df_appended = df.append(df_1)

df_appended

# Create data with different columns from df

data3 = {'customer_id':['6732', '767', '890', '635'], 
        'month':['Jan-20', 'Jan-20', 'Jan-20', 'Jan-20'], 
        'days_purchased_on':[0, 13, 3, 4]} 

# Convert to dataframe

df_2 = pd.DataFrame(data3,index=[0, 1, 2, 3])

df_2

# Create a new dataset combining df and df_2

df_append_test = df.append(df_2)

df_append_test

#### 4. Merge data

# Merge df and df_2 using customer_id as a key 

df_merged = df.merge(df_2, on = ['customer_id'])

df_merged

# Merge df and df_2 using customer_id and month as a keys 

df_merged = df.merge(df_2, on = ['customer_id', 'month'])

df_merged

# Merge df and df_2 using customer_id and month as a keys, add a merge flag

df_merged = df.merge(df_2, on = ['customer_id', 'month'], indicator = True)

df_merged

df_merged['_merge'].value_counts()

# Test merge without overwriting

pd.merge(df,df_2, on = ['customer_id', 'month'], indicator = True)
