==> 02 Scripts/5.0 merging data and exporting.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# loading ords_prior data set into jupythor
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# loading checked_orders data set as df_ords
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

df_ords.head()

df_ords_prior.head(10)

# merging the two data sets on 'order-id'
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# exporting merged df to pkl format
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
