==> 02 Scripts/5.1 merging data and exporting.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# loading orders_products_combined dataset into jupytor
df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# dropping the _merge column as it is of no use here
df_orders_products_combined.drop(columns = '_merge', inplace = True)

# checking shape of imported data set
df_orders_products_combined.shape

df_orders_products_combined.head()

# loading products_checked dataset into jupytor
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

df_prods.shape

df_prods.head()

# merging the 2 data frames on 'product_id' as it appears in both of data sets
df_merged_large = df_orders_products_combined.merge(df_prods, on = 'product_id', indicator = True)

# cheking shape of newly merged data frame 
df_merged_large.shape

# exporting the dataframwe
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))
