==> 02 Scripts/6.0 Deriving New Variable.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# loading orders_products_combined dataset into jupytor
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# slicing a part of data frame 
df = df_ords_prods_merge[:1000000]

df.head()

# a function to flag different price categories
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# calling the funtion
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

### # Use loc() method on df subset to apply price range function.

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 


df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

### Running on entire data frame

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

# checking for the price_range_loc column in df_ords_prods_merge head 
df_ords_prods_merge.head()

### Busiest day of week

# which day most orders take place
df_ords_prods_merge['orders_day_of_week'].value_counts()

#### 0 stands for saturdays which means saturday is the busiest day for instacart followed by sunday and wednesday is the slowest day of week

# creating an empty list 'result' to record entry
result = []

# the for loop will go through each value in orders_day_of_week and
# compare its value with what you know are the busiest and slowest days, and assign it the corresponding string value.

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

#  create a new column within  ords_prods_merge dataframe and set it equal to result
df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

### finding two busiest days and 2 slowest days of week

df_ords_prods_merge['orders_day_of_week'].value_counts()

#### 0 and 1 are the busiest days and 3,4 are the slowest days at insta cart

# creating an empty list 'result' to record entry
result_2 = []

# the for loop will go through each value in orders_day_of_week and
# compare its value with the busiest and slowest days, and assign it the corresponding string value.

for value in df_ords_prods_merge["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest days")
  elif value == 4 or value == 3:
    result_2.append("Least busy days")
  else:
    result_2.append("Regularly busy days")

result_2

#  create a new column within  ords_prods_merge dataframe and set it equal to result_2
df_ords_prods_merge['busiest_days'] = result_2

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

# checking head of df to check the new added column
df_ords_prods_merge.head()

### checking the accuracy of the column
#### as 0 and 1 makes the busiest days category adding up them makes 11864412 which is accurate
#### as 3 and 4 makes the least busy days category adding up them makes 7624336 which is accurate

df_ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

#### as they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders the hours should be devided in three segments [10, 11, 14, 15, 13, 12, 16, 9] as the Most orders, [17, 8, 18, 19, 20, 7, 21, 22] as the avergae orders and [23, 6, 0, 1, 5, 2, 4, 3] as fewest orders 

# creating an empty list 'result' to record entry
result_3 = []

# the for loop will go through each value in order_hour_of_day and
# compare its value with the most orders , average orders and fewest order and assign it the corresponding string value.

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9] :
    result_3.append("Most orders")
  elif value in [17, 8, 18, 19, 20, 7, 21, 22]:
    result_3.append("Average orders ")
  else:
    result_3.append("Fewest orders")

result_3

# addign a new column in the df and assign the result_3 to better use the new column
df_ords_prods_merge['busiest_period_of_day'] = result_3

# Printing frequencyy of the column
df_ords_prods_merge['busiest_period_of_day'].value_counts( dropna = False)

### checking the accuracy of the column
#### as [10, 11, 14, 15, 13, 12, 16, 9] makes the busiest hour category adding up them makes 21118071 which is accurate


### exporting the updating data frame

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_derived.pkl'))
