==> 02 Scripts/7.1 Grouping Data and Aggregation part 2.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# loading orders_products_combined dataset into jupytor
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_01.pkl'))

df_ords_prods_merge.shape

df_ords_prods_merge.head()

### A spending flag for each user based on the average price:

# creating a new column average_order_price to store the avergae order price of order a user has placed
df_ords_prods_merge['average_order_price'] = df_ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# checking the head 50 rows to verify the result of new column
df_ords_prods_merge[['user_id', 'order_number', 'prices', 'average_order_price']].head(50)

# if the avg order price by customer is lesser than 10 then the customer will be flagged as 'low spender' 
df_ords_prods_merge.loc[df_ords_prods_merge['average_order_price'] < 10, 'spender_flag'] = 'Low spender'

# if the avg order price by customer is more or equal than 10 then the customer will be flagged as 'High spender' 
df_ords_prods_merge.loc[df_ords_prods_merge['average_order_price'] >= 10, 'spender_flag'] = 'High spender'

df_ords_prods_merge['spender_flag'].value_counts( dropna = False)

# checking the head to verify the results
df_ords_prods_merge[['user_id', 'prices', 'average_order_price', 'spender_flag']].head(10)

### frequent versus non-frequent customers

# creating a new column customer_frequency to store the median days since prior order of user
df_ords_prods_merge['customer_frequency'] = df_ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

df_ords_prods_merge['customer_frequency'].head(25)

# if the customer_frequency by customer s higher than 20, then the customer should be labeled a “Non-frequent customer.”
df_ords_prods_merge.loc[df_ords_prods_merge['customer_frequency'] > 20, 'order_frequency_flag'] = 'Non-frequent customerr'

# If the customer_frequency is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
df_ords_prods_merge.loc[(df_ords_prods_merge['customer_frequency'] <= 20) & (df_ords_prods_merge['customer_frequency'] > 10), 'order_frequency_flag'] = 'Regular customer'

# If the customer_frequency is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”
df_ords_prods_merge.loc[df_ords_prods_merge['customer_frequency'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

# checking the head to verify the results
df_ords_prods_merge[['user_id', 'customer_frequency', 'order_frequency_flag']].head(20)

df_ords_prods_merge[ 'order_frequency_flag'].value_counts( dropna = False)

#### the NaN here show that there were 5 customers who have only bought once from Instacart

# exporting the df
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_02.pkl'))
