==> 02 Scripts/8.0 Into to Data Visualization with python part 1.ipynb <==
# importing libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

# loading orders_products_combined dataset into jupytor
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_02.pkl'))

# loading customers dataset into jupytor
df_customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

df_customers.shape

df_customers.head(30)

# to Contains a space, which can cause issues when referencing the column programmatically.
# First Name Contains a space, which can cause issues when referencing the column programmatically.
# Surnam Typo in the name (Surnam instead of Surname).
# STATE Uses all uppercase, which is inconsistent with other column names.
# n_dependants Mostly clear, but the n_ prefix might be confusing for some users. Clarify it to number_of_dependents for readability

# renaming columns in customer df
df_customers.rename(columns={
    'First Name': 'first_name',
    'Surnam': 'last_name',
    'Gender' : 'gender',
    'n_dependants' : 'number_of_dependants',
    'STATE': 'state',
    'Age' : 'age',
    'fam_status' : 'family_status'
}, inplace=True)


df_customers.head()

df_customers.describe()

# finding missing values in the data set
df_customers.isnull().sum()

df_customers[df_customers['first_name'].isnull() == True]

### Having NaN values in the firstname column can present issues depending on how the data is processed or analyzed.
### Fill Missing Values: Replace NaN with a placeholder value or an empty string if the firstname column is non-critical:

df_customers['first_name'].fillna('Unknown', inplace=True)

# rerunning the code to verify
df_customers.isnull().sum()

### checking duplicates in the data set

df_customers.duplicated().value_counts()

# datatypes of columns
df_customers.dtypes

# if the data contains any mixed type data 
for col in df_customers.columns.tolist():
  weird = (df_customers[[col]].applymap(type) != df_customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_customers[weird]) > 0:
    print (col)


### so far the df_customers is all good

df_ords_prods_merge.dtypes

# dropping the columns '_merge'
df_ords_prods_merge.drop(columns = '_merge', inplace = True)

### merging the 2 data sets

# Merge the two dataframes using the column 'user_id' as the shared column with the default join (inner)
custs_ords_prods = df_ords_prods_merge.merge(df_customers, on = 'user_id')

custs_ords_prods.shape

custs_ords_prods.head()

# exporting the new data frame
custs_ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'custs_ords_prods.pkl'))
