==> 02 Scripts/9.0 part 1.ipynb <==
# 4.10 Instacart Final Task

## Contents

### Importing Libraries and Dataframes
### Create regional segmentation of data
### Determine whether there’s a difference in spending habits between the different U.S. regions
### Create an exclusion flag for low-activity customers
### Exporting the data set

# import Libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# creating a path to load data fames to python 
path = r"C:\Users\pshan\Downloads\InstaCart Basket Analysis"

ords_prods_custs  = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'custs_ords_prods.pkl'))

ords_prods_custs.shape


# Assign regions using .loc
ords_prods_custs.loc[ords_prods_custs['state'].isin([
    'Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 
    'Connecticut', 'New York', 'Pennsylvania', 'New Jersey'
]), 'region'] = 'Northeast'

ords_prods_custs.loc[ords_prods_custs['state'].isin([
    'Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 
    'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri'
]), 'region'] = 'Midwest'

ords_prods_custs.loc[ords_prods_custs['state'].isin([
    'Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 
    'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 
    'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 
    'Louisiana'
]), 'region'] = 'South'

ords_prods_custs.loc[ords_prods_custs['state'].isin([
    'Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 
    'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 
    'California', 'Hawaii'
]), 'region'] = 'West'



# to verify the results
ords_prods_custs[['state','region']].tail(20)

### Determine whether there’s a difference in spending habits between the different U.S. regions

# Crosstab Region to spending flag
crosstab = pd.crosstab(ords_prods_custs['region'], ords_prods_custs['income'], dropna = False)

crosstab.to_clipboard()

###  Creating an exclusion flag for low-activity customers (customers with less than 5 orders) 

# Create a flag for customers with less than 5 orders
ords_prods_custs.loc[ords_prods_custs['max_order'] < 5, 'activity_flag'] = 'low-activity customer'
ords_prods_custs.loc[ords_prods_custs['max_order'] >= 5, 'activity_flag'] = 'high-activity customer'

# verifying
ords_prods_custs['activity_flag'].value_counts(dropna = False)

# exporting the data frame
ords_prods_custs.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_custs_1.pkl'))

# Exclude low-activity customers from the dataframe
ords_prods_custs_excluded = ords_prods_custs[ords_prods_custs['activity_flag'] != 'low-activity customer']

# Exporting the new dataframe
ords_prods_custs_excluded.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_custs_excluded.pkl'))
